Move frontend tests into individual files.
